+++ test/obsidian/workspace_spec.lua
@@ -34,13 +34,14 @@ describe("Workspace", function()
 
   it("should be able to retrieve the default workspace", function()
     local ws = workspace.get_default_workspace(opts.workspaces)
-    assert.is_not(ws, nil)
+    assert(ws)
     assert.equals(opts.workspaces[1].name, ws.name)
     assert.equals(opts.workspaces[1].path, ws.path)
   end)
 
-  it("should initialize workspace from cwd", function()
+  it("should initialize from cwd", function()
     local ws = workspace.get_workspace_from_cwd(opts.workspaces)
+    assert(ws)
     assert.equals(opts.workspaces[3].name, ws.name)
     assert.equals(opts.workspaces[3].path, ws.path)
   end)
@@ -53,6 +54,7 @@ describe("Workspace", function()
     assert.equals(opts.workspaces[3].path, ws.path)
     opts.detect_cwd = old_cwd
   end)
+
   it("should return default workspace when detect_cwd is false", function()
     local old_cwd = opts.detect_cwd
     opts.detect_cwd = false